[backend] authorize member access by client id

+++ backend/api/widgets/delete.go
@@ -17,7 +17,9 @@ import (
 func Delete(ctx *fasthttp.RequestCtx) {
 	var widget model.Widget
 	if ok := api.FindByID(ctx.UserValue("id"), &widget, ctx); ok && widget.ValidID() {
-		tx := db.Delete(&widget)
-		api.SuccessJSON(ctx, map[string]interface{}{"success": (tx.Error == nil), "error": tx.Error})
+		if deletable := api.CompareClientID(ctx, widget.ClientID, "unprocessible"); deletable {
+			tx := db.Delete(&widget)
+			api.SuccessJSON(ctx, map[string]interface{}{"success": (tx.Error == nil), "error": tx.Error})
+		}
 	}
 }
